Move logic into AudioAnalyzer class
