+++ packages/audio-input/src/index.ts
@@ -2,6 +2,7 @@ export * from './AudioInput'
 export * from './AudioInputPanel'
 export * from './AudioDebugPanel'
 export * from './AudioDeviceManager'
+export * from './AudioAnalyzer'
 
 // Export test utilities
 export { getAudioDiagnostics, testAudioInputCapture } from './AudioTestUtils'
